requirements.txt
matplotlib==2.2.3
numpy==1.18.5
opencv_python==4.5.5.62
pandas==0.23.4
Pillow==9.1.0
prettytable==3.2.0
torchvision==0.9.0
tqdm==4.26.0
